ingestion/mqtt_hfp_ingest/requirements.txt
paho-mqtt
psycopg2-binary
SQLAlchemy
python-dotenv
